requirements.txt
streamlit
python-dotenv
pandas
altair
psycopg[binary]